+++ wirego_plugin/packet-wirego.c
@@ -27,32 +27,17 @@
 #include "plugin-loader.h"
 #include "packet-wirego.h"
 #include "preferences.h"
+#include "dissect.h"
+#include "helpers.h"
 
 void proto_register_wirego(void);
 void proto_reg_handoff_wirego(void);
-static int dissect_wirego(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree _U_, void *data _U_);
-void register_preferences_menu(void);
-enum ftenum field_value_type_to_ws(int vtype);
-field_display_e field_display_type_to_ws(int dtype);
 int get_wireshark_field_id_from_wirego_field_id(int wirego_field_id);
-char * compile_network_stack(packet_info *pinfo);
 void extract_adresses_from_packet_info(packet_info *pinfo, char *src, char *dst);
 
-static int proto_wirego = -1;
-
-//WireGo's subtree
-static int ett_wirego  = -1;
 
 
 
-//Map our go plugin internal field identifiers to the ones provided by Wireshark
-typedef struct {
-  int wirego_field_id;
-  int wireshark_field_id;
-} field_id_to_plugin_field_id_t;
-
-int fields_count = -1;
-field_id_to_plugin_field_id_t * fields_mapping = NULL;
 
 //Register protocol when plugin is loaded.
 void proto_register_wirego(void) {
@@ -225,125 +210,3 @@ void proto_reg_handoff_wirego(void) {
   }
 
 }
-
-static int
-dissect_wirego(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree, void *data _U_)
-{
-  int pdu_len;
-  char * golang_buff = NULL;
-  char src[255];
-  char dst[255];
-  char * full_layer = NULL;
-  int dissectHandle = -1;
-
-  /*
-    In a more classic Wireshark plugin we would use all the tvb_* accessors here
-    Since processing of the packet is performed in the golang plugin (that's actually the very purpose
-    of this insanity), and since I won't write bindings for the complete wireshark API, we need to push
-    the packet buffer to the plugin.
-
-    We have two options here:
-
-      - use tvb_get_ptr
-      - use tvb_memcpy
-    
-    The get_ptr would be the more obvious, but it is marked as very dangerous. Since this buffer would be pushed 
-    to a golang plugin it could eventually be even more dangerous.
-    Thus we're using tvb_memcpy, which will provide us a dedicated buffer to play with.
-    That's not optimal at all, but we'll start with this.
-  */
-
-
-  if (!tvb || !pinfo)
-    return -1;
-
-  pdu_len = tvb_reported_length(tvb);
-  if (pdu_len <= 0)
-    return 0;
-
-  src[0] = 0x00;
-  dst[0] = 0x00;
-  extract_adresses_from_packet_info(pinfo, src, dst);
-
-
-  full_layer = compile_network_stack(pinfo);
-
-  //Pass everything to the golang plugin
-  golang_buff = (char*) malloc(pdu_len);
-  tvb_memcpy(tvb, golang_buff, 0, pdu_len);
-  dissectHandle = wirego_dissect_packet_cb(pinfo->num, src, dst, full_layer, golang_buff, pdu_len);
-  free(golang_buff);
-  golang_buff = NULL;
-  free(full_layer);
-  full_layer = NULL;
-
-  if (dissectHandle == -1) {
-    //col_set_str will keep a pointer to the given value
-    //while col_add_str will duplicate
-    col_set_str(pinfo->cinfo, COL_PROTOCOL, "Wirego plugin failed.");
-    col_set_str(pinfo->cinfo, COL_INFO, "Wirego plugin failed.");
-    return -1;
-  }
-
-  //Analyse plugin results
-
-  //Flag protocol name
-  col_set_str(pinfo->cinfo, COL_PROTOCOL, wirego_result_get_protocol_cb(dissectHandle));
-
-  //Fill "info" column
-  col_set_str(pinfo->cinfo, COL_INFO, wirego_result_get_info_cb(dissectHandle));
-
-  //During the first pass, tree can eventually be NULL
-  //Wireshark does not ask the plugin to fill detailed structures
-  if (tree) {
-    //How many custom fields did the plugin return?
-    int result_fields_count = wirego_result_get_fields_count_cb(dissectHandle);
-    if (result_fields_count != 0) {
-      
-      //Add a subtree on this packet
-      proto_item *ti = proto_tree_add_item(tree, proto_wirego, tvb, 0, -1, ENC_BIG_ENDIAN);
-      if (!ti) goto DONE;
-      proto_tree *wirego_tree = proto_item_add_subtree(ti, ett_wirego);
-      if (!wirego_tree)  goto DONE;
-
-      //Process all custom fields
-      for (int i = 0; i < result_fields_count; i++) {
-        int wireshark_field_id = -1;
-        int wirego_field_id;
-        int offset;
-        int length;
-        //Ask plugin for result
-        wirego_result_get_field_cb(dissectHandle, i, &wirego_field_id, &offset, &length);
-        //Convert plugin field id to wireshark id
-        wireshark_field_id = get_wireshark_field_id_from_wirego_field_id(wirego_field_id);
-        //Add tree entry
-        if (wireshark_field_id != -1) {
-          proto_item *sub = proto_tree_add_item(wirego_tree, wireshark_field_id, tvb, offset, length, ENC_BIG_ENDIAN);
-          proto_tree *subsub = proto_item_add_subtree(sub, ett_wirego);
-          proto_tree_add_item(subsub, wireshark_field_id, tvb, offset, length, ENC_BIG_ENDIAN);
-
-        }
-      }    
-    }
-  }
-DONE:
-  wirego_result_release_cb(dissectHandle);
-  return tvb_captured_length(tvb);
-}
-
-
-
-
-
-
-//Convert a field id, as provided by the Golang plugin to a Wireshark filed id,
-//as returned by wireshark backend during declaration
-int get_wireshark_field_id_from_wirego_field_id(int wirego_field_id) {
-  for (int idx = 0; idx < fields_count; idx++) {
-    if (fields_mapping[idx].wirego_field_id == wirego_field_id) {
-      return fields_mapping[idx].wireshark_field_id;
-    }
-  }
-  return -1;  
-}
-
